+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _Not yet on NuGet..._
 * Controls: Improve axis lock behavior when dragging the mouse on a control (#3498) @BrianAtZetica
 * Heatmap: Added `Opacity` and `AlphaMap` properties to enhance transparency customization (#3499, #3349) @BrianAtZetica
 * Heatmap: Intensity values that are `double.NaN` are now displayed as transparent cells (#3499, #3349) @BrianAtZetica
+* Text: Added an `OffsetX` and `OffsetY` properties for adjusting text position in pixel units (#3506) @jamaa
 
 ## ScottPlot 5.0.22
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-03-16_